moved helper functions to helper, added docstrings
